pyproject.toml
[project]
name = "comunicarlos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.122.0",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.15.5",
    "python-dateutil>=2.9.0.post0",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn>=0.38.0",
]
